% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dis_fac.R
\name{TablaAnova}
\alias{TablaAnova}
\title{Dise単o factorial con dos factores.}
\usage{
TablaAnova(respuesta, factorA, factorB, data)
}
\arguments{
\item{respuesta}{(Columna de datos) Nuestra variable dependiente o variable respuesta.}

\item{factorA}{(Columna de datos) El primer factor de nuestro dise単o factorial.}

\item{factorB}{(Columna de datos) El segundo factor de nuestro dise単o factorial:}
}
\value{
Devuelve una tabla ANOVA con la fuente de variacion, la suma de cuadrados, los grados de libertad, el cuadrado medio y el F valor.
}
\description{
Crea un dise単o factorial de dos factores.
}
\examples{
\dontrun{
library(readr)
Libro1 <- read_csv("C:/Users/donpo/OneDrive/Escritorio/Libro1.csv")

library(DisFact)

data <- data.frame(Libro1)

TablaAnova(respuesta = "hr", factorA = "temp", factorB = "mat", data = data)
}
}
